Create fundamentals.py
Creating a python project can be a daunting task to imagine as a beginner. 
Before we reach that stage we want to have the best foundation possible. 
To make that happen we should practice as much as possible. Patience and and persistence is key!

If you are a more experienced  user of either python or Github, I would love any constructive feed back. Thanks!